neural-network-db-project/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
python-multipart==0.0.6
pydantic==2.5.0
aiofiles==23.2.1
Pillow==10.0.1
python-jose[cryptography]==3.3.0
